cloud/orchestrator/requirements.txt
pytest
sense-o-api